CASMPET-3815: sysmgmt-health node-exporter lower CPU usage
The node-exporter pods were getting throttled by K8s CPU limiter.
Investigation found that on w001 of one system (sif) the following
collectors were taking a long time to run:

* netclass (5 seconds)
* ntp (1 second) -- this was also always failing
* qdisc (115 seconds)

The node-exporter on the node was also mostly unresponsive
(requests to /metrics failed after a long time).

The proposed fix is to disable these collectors.

In addition, the GOMAXPROCS environment variable is set to 1. By
default the go runtime will allocate 1 thread per CPU, which when
we've got 64 CPUs makes for a large number of threads which
creates a lot of overhead. Since we're limiting the number of CPUs
used to 4 in K8s the original setting isn't doing any good. There's
a bug[1] and PR to make this change upstream.

[1] https://github.com/prometheus/node_exporter

The chart being used doesn't allow passing in extra environment
variables so this is done with a chart hook. The kubectl image
used is one that's used in other charts.

JIRAs will be opened to investigate re-enabling these collectors.