CASMPET-4912: Fix collecting metrics for / on K8s NCNs
node-exporter wasn't collecting any metrics for the / filesystem
because on the K8s NCNs its FSTYPE is 'overlay':

```
ncn-m001:~ # findmnt /
TARGET SOURCE        FSTYPE  OPTIONS
/      LiveOS_rootfs overlay rw,relatime,lowerdir=/run/rootfsbase,upperdir=/run/overlayfs,workdir=/run/ovlwork
```

and the filesystem collector is configured to ignore overlay fs's. The
'/' filesystem wasn't an overlay FS originally and overlay fs's are
used for the containers so these were ignored.

This was causing there to be no alert when the `/` filesystem fills up
due to kubelet spewing messages into the logs.

The fix is to not ignore overlay fs's and make sure that the container
fs's are ignored.