CASMPET-4912: Add back excluded containerd mount point
When https://github.com/Cray-HPE/cray-sysmgmt-health/pull/27/files was
made, it removed part of the changes done for
https://github.com/Cray-HPE/cray-sysmgmt-health/commit/68e37942257034ca2c73ea79194ec584d0dddf3b
to "fix the mount points to ignore that node-exporter sees as an
errored devices". This was unintentional and due to developing using
the old charts that are in vshasta. This change adds that back in.