CASMPET-4467: Add dynamic monitoring (for Istio)
This adds a Pod monitor that dynamically finds Pods to monitor, based
on annotations on the Pod.

This was copied from the configuration that's used in the Prometheus
instance that's deployed by Istio. It was already doing this dynamic
monitoring to find istio-proxy sidecars and istiod.

The Istio Prometheus instance only has the following targets:
* kubernetes-apiservers
* kubernetes-cadvisor
* kubernetes-nodes
* kubernetes-pods
* kubernetes-service-endpoints

The sysmgmt-health Prometheus already has monitors for kubernetes
(apiservers, cadvisor, and nodes), so I didn't add new monitors for
those.

The kubernetes-service-endpoints created a dynamic monitor for
Services. Because we already have several existing service monitors
and services that have the annotations used by the new dynamic monitor
there wound up being several services that were being scraped twice
with different labels. I investigated some of these and wasn't sure
that this wouldn't cause problems with the alerts and grafana
dashboards so decided to leave it out when it looked like we wouldn't
be missing anything. istio is covered by the kubernetes-pods monitor.

For the kubernetes-pods PodMonitor, I got the configuration from
Istio's Prometheus. It's in the `istio-system/prometheus` ConfigMap:

```
...
- job_name: 'kubernetes-pods'
  kubernetes_sd_configs:
  - role: pod
  relabel_configs:  # If first two labels are present, pod should be scraped  by the istio-secure job.
  - source_labels: [__meta_kubernetes_pod_annotation_prometheus_io_scrape]
    action: keep
    regex: true
  - source_labels: [__meta_kubernetes_pod_annotation_prometheus_io_path]
    action: replace
    target_label: __metrics_path__
    regex: (.+)
  - source_labels: [__address__, __meta_kubernetes_pod_annotation_prometheus_io_port]
    action: replace
    regex: ([^:]+)(?::\d+)?;(\d+)
    replacement: $1:$2
    target_label: __address__
  - action: labelmap
    regex: __meta_kubernetes_pod_label_(.+)
  - source_labels: [__meta_kubernetes_namespace]
    action: replace
    target_label: namespace
  - source_labels: [__meta_kubernetes_pod_name]
    action: replace
    target_label: pod_name
...
```

Then I just converted this to a PodMonitor.

The istio Prometheus doesn't define any alerts.

When this is in place the Prometheus instance deployed by Istio can be
removed. We need to do something with the Istio Prometheus because the
Istio operator will no longer support deploying Prometheus in newer
releases.

Grafana is reconfigured to point the `istio-monitoring` data source
from the istio prometheus to the sysmgmt prometheus since this now
has the istio metrics and the istio Prometheus is going away.