CASMPET-4467: Add dynamic monitoring (for Istio)
This adds a Pod and Service monitor that dynamically finds Pods or
Services to monitor, based on annotations on the Pod or Service. See
the README.

This was copied from the configuration that's used in the Prometheus
instance that's deployed by Istio. It was already doing this dynamic
monitoring to find istio-proxy sidecars and istiod, but it was also
catching several other Services that configure the annotations and we
were getting for free but wouldn't be using anyways.

The Istio Prometheus instance only has the following targets:
* kubernetes-apiservers
* kubernetes-cadvisor
* kubernetes-nodes
* kubernetes-pods
* kubernetes-service-endpoints

The sysmgmt-health Prometheus already has monitors for kubernetes
(apiservers, cadvisor, and nodes), so I didn't change that.

For the kubernetes-pods PodMonitor, I got the configuration from
Istio's Prometheus. It's in the ConfigMap:

```
...
- job_name: 'kubernetes-nodes'
  scheme: https
  tls_config:
    ca_file: /var/run/secrets/kubernetes.io/serviceaccount/ca.crt
  bearer_token_file: /var/run/secrets/kubernetes.io/serviceaccount/token
  kubernetes_sd_configs:
  - role: node
  relabel_configs:
  - action: labelmap
    regex: __meta_kubernetes_node_label_(.+)
  - target_label: __address__
    replacement: kubernetes.default.svc:443
  - source_labels: [__meta_kubernetes_node_name]
    regex: (.+)
    target_label: __metrics_path__
    replacement: /api/v1/nodes/${1}/proxy/metrics
...
```

Then I just converted this to a PodMonitor.

For the kubernetes-service-endpoints ServiceMonitor, that's also in the ConfigMap:

```
- job_name: 'kubernetes-service-endpoints'
  kubernetes_sd_configs:
  - role: endpoints
  relabel_configs:
  - source_labels: [__meta_kubernetes_service_annotation_prometheus_io_scrape]
    action: keep
    regex: true
  - source_labels: [__meta_kubernetes_service_annotation_prometheus_io_scheme]
    action: replace
    target_label: __scheme__
    regex: (https?)
  - source_labels: [__meta_kubernetes_service_annotation_prometheus_io_path]
    action: replace
    target_label: __metrics_path__
    regex: (.+)
  - source_labels: [__address__, __meta_kubernetes_service_annotation_prometheus_io_port]
    action: replace
    target_label: __address__
    regex: ([^:]+)(?::\d+)?;(\d+)
    replacement: $1:$2
  - action: labelmap
    regex: __meta_kubernetes_service_label_(.+)
  - source_labels: [__meta_kubernetes_namespace]
    action: replace
    target_label: kubernetes_namespace
  - source_labels: [__meta_kubernetes_service_name]
    action: replace
    target_label: kubernetes_name
```

Then I just converted this to a ServiceMonitor.

The istio Prometheus doesn't define any alerts.

When this is in place the Prometheus instance deployed by Istio can be
removed. We need to do something with the Istio Prometheus because the
Istio operator will no longer support deploying Prometheus in newer
releases.