CASMPET-5038: Update image refs in charts
Followed the objectives in CASM-2670 [0].

All of the images being used are now from algol60 artifactory. The
one exception is smartmon which is now using the community image. The
image built from container-images had a new vulnerability and a
rebuild didn't fix the problem. That image is no longer built by
container-images after some investigation of how it was being built
and we'll use the community image rather than rebuild it.

The community smartmon image also has the vulnerability but there's
no fix for it yet. I tried to configure snyk to ignore the
vulnerability but that's not working for some reason. This is why
scan-images is set to `false` in the workflows. A JIRA will be opened
to investigate why ignoring vulnerabilities isn't working.

scan-charts is set to `false` in the workflows because this was also
failing. A JIRA will be opened to investigate this.

[0] https://jira-pro.its.hpecorp.net:8443/browse/CASM-2670